% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addrows.R
\name{addrows}
\alias{addrows}
\title{Add rows to a data.frame}
\usage{
addrows(dtf, nrw, top = FALSE)
}
\arguments{
\item{dtf}{data.frame; the object that shall be added to}

\item{nrw}{data.frame; the new row(s) to be added}

\item{top}{logical; should the new rows be added to the top or the bottom (default)?}
}
\description{
An "\code{rbind} for data.frames", sort of.
}
\details{
Can only bind two objects at a time, but will bind data.frames with
non-matching column names and -classes. In such cases the first data.frame will serve as
template.
}
\examples{
dtf <- data.frame(A=letters[1:5], 
                  B=1:5, 
                  C=as.factor(5:1), 
                  D=as.Date(0:4, origin="2000-01-01"),
                  stringsAsFactors=FALSE)

nrw <- data.frame(A=letters[1:5], 
                  B=4:8, 
                  C=5:1, 
                  D=as.Date(5:1, origin="1990-01-01"),
                  stringsAsFactors=FALSE)
str(dtf)

dtf.a <- addrows(dtf, nrw, top=FALSE)
str(dtf.a)

# adding a single row with little concern for data types and column names
b <- type.convert(beaver1[80:90,])
b$activ <- as.logical(b$activ)

addrows(b, data.frame(350, 1200, 37.02, 1))
}
