% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustcolorHSV.R
\name{adjustcolorHSV}
\alias{adjustcolorHSV}
\title{Adjust Colors in One or More Directions Conveniently.}
\usage{
adjustcolorHSV(col, alpha.f = 1, h.f = 1, s.f = 1, v.f = 1,
  offset = c(0, 0, 0, 0), transform = diag(c(h.f, s.f, v.f, alpha.f)))
}
\arguments{
\item{col}{vector of colors, in any format that col2rgb() accepts}

\item{alpha.f, h.f, s.f, v.f}{factors scaling the opacity, hue, saturation 
and value of \code{col}}

\item{offset}{a length 4 numeric vector specifying the linear offset applied
to the \emph{hue}, \emph{saturation}, \emph{value} and \emph{alpha} values}

\item{transform}{a 4x4 diagonal matrix specifying the scaling applied to the
\emph{hue}, \emph{saturation}, \emph{value} and \emph{alpha} values}
}
\value{
A character vector the same length as \code{col} contaning color data in standard
hexadeximal RGBA format.
}
\description{
Adjust or modify a vector of colors by "turning knobs" on one or more coordinates
in (h,s,v,\eqn{\alpha}) space, typically by up or down scaling them.
}
\details{
Essentially an HSV version of the RGB-based \code{\link{adjustcolor}}. One
important distinction is that the \code{h.f} value wraps around to fit the [0, 1]
range, rather than simply "clamping" it between 0 and 1.
}
\examples{
# Halve the saturation and value of the default palette colours
plot(2:8, cex=5, lwd=4, pch=21, bg=2:8, 
  col=adjustcolorHSV(2:8, s.f=0.5, v.f=0.6))

# Offset the hue of the default palette colours by 0.5, inverting the colours
plot(2:8, cex=5, lwd=4, pch=21, bg=2:8, 
  col=adjustcolorHSV(2:8, offset=c(0.5, 0, 0, 0)))

}
