% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averaged shifted histograms.R
\name{ahist}
\alias{ahist}
\title{Average shifted histogram}
\usage{
ahist(x, n.breaks = nclass.FD(x), n.shifts = 3, type = c("histogram",
  "polygon", "line", "table"), freq = FALSE, plot = TRUE, add = FALSE,
  ...)
}
\arguments{
\item{x}{a vector of values for which the histogram is desired}

\item{n.breaks}{an integer giving the number of bins to be used}

\item{n.shifts}{an integer giving the number of shifts to be performed}

\item{type}{if \code{plot=TRUE}, the type of plot to be used}

\item{freq}{should frequency counts be used, or density (default)}

\item{plot}{logical; if \code{TRUE} (default), a graphical output will be 
returned}

\item{add}{logical; if \code{TRUE} the plot will be added to the current plot}

\item{...}{further graphical parameters to \code{ymse::plot.histogram}, 
\code{polygon}, or \code{lines}}
}
\value{
an object of class \code{"histogram"}
}
\description{
Create a smoothed histogram by averaging several histograms shifted by fractions 
of a bin-width
}
\examples{
set.seed(1)
n <- 6

x <- sample(sample(0:20, 8), 6*n, replace=TRUE) + rnorm(6*n, -8, 0.5)
x <- c(x, rgamma(5*n, 3, 0.5), rnorm(4*n, 15, 2))
x <- round(x*5)/5

hist(x, freq=FALSE, breaks="FD", col="lightblue")
ahist(x, type="hist", border=2, col=7, freq=FALSE, lwd=2)
ahist(x, type="poly", border=2, col=7, freq=FALSE, lwd=2)
ahist(x, type="line", col=2, freq=FALSE, lwd=2)
ahist(x, type="table", col=2, freq=FALSE, lwd=2)
ahist(x, plot=FALSE)
}
