% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_char.R
\name{align_char}
\alias{align_char}
\title{Align character strings}
\usage{
align_char(x, pattern = ".", ..., lpad = " ", rpad = " ")
}
\arguments{
\item{x}{numeric or character vector}

\item{pattern}{pattern, passed to \code{regexpr}, whose match the 
character strings will be aligned by}

\item{...}{further arguments passed to \code{regexpr}}

\item{lpad, rpad}{character strings used for padding. Repeated to length}
}
\description{
Align character strings
}
\examples{
x <- c("Tom und Jerry", "Abbott og Costello", "Milo och Stich", "et alii")
cat(align_char(x, pat="[[:alpha:]]"), sep="\\n")
cat(align_char(x, pat=" "), sep="\\n")
cat(align_char(x, pat=" [A-Z]"), sep="\\n")
cat(align_char(x, pat=" [a-z]"), sep="\\n")
cat(align_char(x, pat="t", ignore.case=TRUE), sep="\\n")
cat(align_char(x, pat="x"), sep="\\n")
}
\seealso{
\code{\link{align_num}}, for alignment more aimed at numeric strings
}
