% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armodel.R
\name{arfilter}
\alias{arfilter}
\title{AR filter}
\usage{
arfilter(x, mod, x.mean = mod$x.mean, init = "focb")
}
\arguments{
\item{x}{a time series}

\item{mod}{an AR model}

\item{x.mean}{the mean used. By default the mean of the original model. Set to zero for
no demeaning}
}
\description{
Filter a time series using AR coefficients
}
\examples{
set.seed(1)
arap <- ar(AirPassengers)
spec.ar(arap)
spec.pgram(arfilter(rnorm(10000), arap), span=21, na.action=na.omit)

arm <- armodel(c(1.3, -0.4))
spec.ar(arm)
plot(x <- rnorm(200), type="l")
lines(scale(arfilter(x, arm), center=FALSE), col="red", lwd=2)
}
\seealso{
\code{\link{armodel}}
}
