% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armodel.R
\name{arfit}
\alias{arfit}
\title{AR model fit}
\usage{
arfit(x, mod, x.mean = mod$x.mean)
}
\arguments{
\item{x}{a time series}

\item{mod}{an AR model}

\item{x.mean}{the mean used. By default the mean of the original model.
Set to zero for no demeaning}
}
\description{
Fit a specified AR model to a univariate time series
}
\examples{
set.seed(1)
x <- runif(50) + sin(1:50/10)
plot(x); lines(arfilter(x, armodel(c(1.5, -0.5, 0.5)), x.mean=mean(x)))
}
\seealso{
\code{\link{armodel}} for examples
}
