% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armodel.R
\name{arimpulse}
\alias{arimpulse}
\title{Impulse response of an AR model}
\usage{
arimpulse(mod, pulse = 1, n.ahead = 20, plot = TRUE, ...)
}
\arguments{
\item{mod}{an AR model}

\item{pulse}{numeric vector; the initial pulse. Magnitude is added to the model mean}

\item{n.ahead}{the length of the computed response}

\item{plot}{logical; sgould the result be plotted?}

\item{...}{further arguments to \code{plot}}
}
\description{
Get and plot the impulse response of an AR model
}
\seealso{
\code{\link{armodel}} for examples
}
