% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.array.list.R
\name{as.array.list}
\alias{as.array.list}
\title{Coerce a list to an array}
\usage{
\method{as.array}{list}(x, ...)
}
\arguments{
\item{x}{a list of equal sized data.frames or matrices}

\item{...}{(not used)}
}
\value{
A list of length \eqn{l} with elements of \eqn{m} rows and \eqn{n} 
columns wix result in an \eqn{m \times n \times l}{m × n × l} array.
}
\description{
Coerce a list consisting of data.frames or matrices of equal size to a 3d array
}
\examples{
df1 <- data.frame(x=c(1, 2, 3), y=c(2, 3, 4), z=c(3, 4, 5))
df2 <- data.frame(x=c(4, 2, 3), y=c(2, 5, 4), z=c(3, 4, 6))
df3 <- data.frame(x=c(1, 4, 2), y=c(3, 3, 8), z=c(4, 3, 5))

l <- list(df1, df2, df3)

as.array(l)

as.array(speedskate)
}
