% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmented_median.R
\name{aug_median}
\alias{aug_median}
\title{Centre weighted mean}
\usage{
aug_median(x, p = 1, rank = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{p}{positive numeric narrowness of the weight. \code{1} gives triangular 
weighting. Higher values gives narrower weights, approching meadian, lower values 
gives broader weights, approaching arithmetic mean}

\item{rank}{logical. Should should ranks or numeric values determine relative 
weights?}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
Offering a continuous link between median and arithmetic mean
}
\details{
A weighted arithmetic mean is calculated over the input vector, where
most weight is given to the median value(s), and monotonically less towards either
extreme. Faloff depends on \code{p}, with small values resulting in a gentler 
falloff and less difference between minimum and maximum weights.
}
\examples{
x <- c(0, 8, 8, 8, 9)

aug_median(x)

# 0 and 9 are considered equidistant from 8
aug_median(x, rank=TRUE)

# Nearly a point weight placed at the median
aug_median(x, 100)
median(x)

# Nearly uniform weights
aug_median(x, 0.001)
mean(x)
}
