% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{bix}
\alias{bix}
\alias{bix<-}
\title{Bix attributes}
\usage{
bix(d)

bix(d) <- value
}
\arguments{
\item{d}{a \code{"dice"} object}

\item{value}{value to begin index at}
}
\description{
\code{bix} provides access to the bix attribute of a variable. The
first form returns the value of the levels of its argument and the second
sets the attribute.
}
\examples{
d <- dice(6)
d
bix(d)
bix(d) <- 3
d
expand(d)
}
