% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caleidoscope.R
\name{caleidoscope}
\alias{caleidoscope}
\title{Caleidoscopic effect on a matrix}
\usage{
caleidoscope(m, odd = TRUE)
}
\arguments{
\item{m}{a matrix}

\item{odd}{logical; should the resulting matrix have odd dimensions?}
}
\value{
A matrix of either \eqn{2\times}{2×} or \eqn{2\times - 1}{2× - 1} the number
of rows and columns of the input matrix.
}
\description{
Flip a matrix vertically and horizontally before recombining into a new large matrix
}
\details{
Three copies of \code{m} will be made. One flipped horizontally, one flipped
vertically, and one flipped both horizontally and vertically. Then they are recombined
with the original matrix in the upper right corner, and the flipped copies in the
upper left, lower righ and lower left corners, respectively.
}
\examples{
caleidoscope(matrix(1:4, 2), odd=FALSE)

image(caleidoscope(1:9 \%o\% 1:9))

image(caleidoscope(matrix(runif(180*200)^2, 180)), col=rainbow(256, start=0.58))
}
