% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerknit.R
\name{cbapply}
\alias{cbapply}
\title{Apply function to contents of clipboard}
\usage{
cbapply(FUN, ..., collapse = FALSE, write = TRUE, eval = FALSE)
}
\arguments{
\item{FUN}{function to be applied}

\item{...}{optional arguments to \code{FUN}}

\item{collapse}{collapse the lines into a single string separated by newlines}

\item{write}{write the results back to the clipboard}

\item{eval}{parse and evaluate the results}
}
\description{
Read in clipboard contents as lines, apply a function on them, and write
results back to the clipboard
}
\examples{

\dontrun{
# Copy to clipboard
a <- 10
b <- 20
s <- a + b
s
# end

# Run
ev <- cbapply(FUN=function(x) paste(x, "+ 2"), eval=TRUE)
ev; a; b; s

# Clipboard contents changed to
a <- 10 + 2
b <- 20 + 2
s <- a + b + 2
s + 2
# end

# Copy to clipboard
One Two
Three
# end

# Run
cbapply(FUN=toupper, write=FALSE)

# Clipboard contents unchanged
One Two
Three
# end
}
}
