% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_with_ties_tgt_tlt.R
\name{comparison_with_ties}
\alias{comparison_with_ties}
\alias{\%tgt\%}
\alias{tgt}
\alias{\%tlt\%}
\alias{tlt}
\title{Comparison with ties}
\usage{
x \%tgt\% y

tgt(x, y, bias = 0.5)

x \%tlt\% y

tlt(x, y, bias = 0.5)
}
\arguments{
\item{x, y}{numeric values to be compared}

\item{bias}{what bias should be given to ties? 0.5, the default, is 
considered neutral
as it's halfway between 1 and 0 (true and false).}
}
\description{
Compare numeric values, returning an inbetween value for ties
}
\examples{
1:5 \%tlt\% 3
1:5 \%tgt\% 3

c(1, 4, 3, 1) \%tlt\% c(1, 3, 3, 2)
c(1, 4, 3, 1) \%tgt\% c(1, 3, 3, 2)

# Calculate Mann–Whitney U statistic
set.seed(1)
x <- sort(round(runif(20)*13, 1))
y <- sort(round(runif(15)*10, 1))
o <- outer(x, y, "\%tgt\%")

sum(o)
wilcox.test(x, y, exact=FALSE)$statistic


}
\seealso{
\code{\link{Comparison}}, \code{\link{tied_triple_test}}
}
