% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar_and_par.R
\name{default_par}
\alias{default_par}
\title{Default par}
\usage{
default_par()
}
\description{
Sets \code{par} settings to their default values
}
\details{
Default par settings can be retreived by \code{data(.def.par)}. A new default
can be specified by editing \code{def.par} or making a 
\code{def.par <- par(no.readonly=TRUE)} type call.
}
\seealso{
Other par_and_plot_margins_functions: \code{\link{revert_par}},
  \code{\link{set_mar}}
}
