% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{dice}
\alias{dice}
\alias{is.dice}
\alias{as.dice}
\alias{print.dice}
\alias{as.table.dice}
\title{Create, modify or convert from/to dice objects}
\usage{
dice(dval)

is.dice(x, ...)

as.dice(x, ...)

\method{print}{dice}(x, ...)

\method{as.table}{dice}(x, ...)
}
\arguments{
\item{dval}{an integer vector}

\item{x}{an arbitrary \code{R} object}

\item{...}{further arguments passed to methods}
}
\description{
Create, modify or convert from/to dice objects
}
\examples{
# Regular d6 dice
dice(6)

# d4 dice with sides 0, 1, 2, 4
dice(c(0:3))

# d4 dice with two 2s and two 5s
dice(c(2, 2, 5, 5))

}
\seealso{
\code{\link{expand}}, \code{\link{table}}
}
