% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dput2.R
\name{dput2}
\alias{dput2}
\title{Write an Object to console}
\usage{
dput2(x, width = 65, assign = c("front", "end", "none"))
}
\arguments{
\item{x}{r object}

\item{width}{width of column}

\item{assign}{should assignment to object go in front of or at the end of the object,
or none at all?}
}
\description{
Writes an ASCII text representation of an R object to the console
}
\details{
This is similar to the way \code{dput} is used to print ASCII representations
of objects to the console. The differences are that \code{dput2} lets you specify
column width, and assignment to the object is an option.
}
\examples{
xmpl <- faithful[sort(sample(1:nrow(faithful), 50)), ]

dput(xmpl)
cat(deparse(xmpl, width.cutoff=65), sep='\\n')

dput2(xmpl, 65)
}
