% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_effects.R
\name{drop_randfx}
\alias{drop_randfx}
\title{Drop random effects}
\usage{
drop_randfx(form)
}
\arguments{
\item{form}{a formula object}
}
\value{
A formula object
}
\description{
Drop random effects from a mixed effects model formula
}
\details{
\code{form} is divided into its individual terms, any term containg a vertical
bar (\code{|}) is removed, before \code{form} is updated and returned. 
In case \code{form} has no random effect terms, \code{form} is returned 
unmodified. In case all effects are random, only the intercept is retained.
In any case the response variable(s) are kept as is.
}
\examples{
f1 <- Reaction ~ (1 + Days | Subject)
f2 <- Reaction ~ (1 | mygrp/mysubgrp) + (1 | Subject)
f3 <- Reaction ~ x1 + x2 + (1 + Days | Subject)
f4 <- Reaction ~ x1 * x2 + (1 | mygrp/mysubgrp) + (1 | Subject)
f5 <- Reaction ~ x1 + x2

sapply(list(f1, f2, f3, f4, f5), drop_randfx)

}
\seealso{
\code{\link{drop_pattern}}
}
