% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtf_clean.R
\name{dtf_clean}
\alias{dtf_clean}
\title{Data cleanup}
\usage{
dtf_clean(
  x,
  header = TRUE,
  na.strings = c("NA", "N/A"),
  stringsAsFactors = FALSE,
  ...
)
}
\arguments{
\item{x}{a messy table the form of a character string}

\item{header}{does the table include headers? (default TRUE)}

\item{na.strings}{a vector of character strings which will be interpreted
as missing values}

\item{stringsAsFactors}{should strings be read as factors? (default FALSE)}

\item{...}{further arguments passed to \code{read.table}}
}
\description{
Create a data.frame from a messy table
}
\examples{
\dontrun{
x1 <- "
+------------+------+------+----------+--------------------------+
|    Date    | Emp1 | Case | Priority | PriorityCountinLast7days |
+------------+------+------+----------+--------------------------+
| 2018-06-01 | A    | A1   |        0 |                        0 |
| 2018-06-03 | A    | A2   |        0 |                        1 |
| 2018-06-02 | B    | B2   |        0 |                        2 |
| 2018-06-03 | B    | B3   |        0 |                        3 |
+------------+------+------+----------+--------------------------+
"

x2 <- '
------------------------------------------------------------------
|    Date    | Emp1 | Case  | Priority | PriorityCountinLast7days |
------------------------------------------------------------------
| 2018-06-01 | A    | "A 1" |        0 |                        0 |
| 2018-06-03 | A    | "A 2" |        0 |                        1 |
| 2018-06-02 | B    | "B 2" |        0 |                        2 |
| 2018-06-03 | B    | "B 3" |        0 |                        3 |
------------------------------------------------------------------
'

x3 <- "
––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
    Date    | Emp1 | Case | Priority | PriorityCountinLast7days 
––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
 2018-06-01 | A    | A|1  |        0 |                        0 
 2018-06-03 | A    | A|2  |        0 |                        1 
 2018-06-02 | B    | B|2  |        0 |                        2 
 2018-06-03 | B    | B|3  |        0 |                        3 
––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
"

x4 <- "
 Maths | English | Science | History | Class

  0.1  |  0.2    |  0.3    |  0.2    |  Y2

  0.9  |  0.5    |  0.7    |  0.4    |  Y1

  0.2  |  0.4    |  0.6    |  0.2    |  Y2

  0.9  |  0.5    |  0.2    |  0.7    |  Y1
"

x5 <- "
       Season   |   Team  | W | AHWO
-------------------------------------
1  |  2017/2018 |  TeamA  | 2 | 1.75
2  |  2017/2018 |  TeamB  | 1 | 1.85
3  |  2017/2018 |  TeamC  | 1 | 1.70
4  |  2017/2018 |  TeamD  | 0 | 3.10
5  |  2016/2017 |  TeamA  | 1 | 1.49
6  |  2016/2017 |  TeamB  | 3 | 1.51
7  |  2016/2017 |  TeamC  | 2 | 1.90
8  |  2016/2017 |  TeamD  | 0 | N/A 
"

lapply(c(x1, x2, x3, x4), dtf_clean)
}
}
