% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\alias{entropy.table}
\alias{entropy.data.frame}
\alias{entropy.matrix}
\alias{entropy.default}
\title{Information entropy}
\usage{
\method{entropy}{table}(x, base = 2)

\method{entropy}{data.frame}(x, base = 2)

\method{entropy}{matrix}(x, base = 2)

\method{entropy}{default}(x, base = 2)
}
\arguments{
\item{x}{a vector, table, data.frame or matrix. In the case of table, data.frame and matrix
each row is treated as separate tabulated data.}

\item{base}{the log base to be used.}
}
\description{
Computes the information entropy (also called Shannon entropy) of a set of discrete values,
or a tabulated such set.
}
\examples{
entropy(c(5, 5, 4, 4, 2, 3, 5))  # default is unit bits
entropy(c(5, 5, 4, 4, 2, 3, 5), base=exp(1))  # unit nats

entropy(rep(1:4, 1:4), 4)
entropy(rep(1:4, 1), 4)

mtctab <- table(mtcars$cyl, mtcars$carb)
entropy(mtctab, 6)

xx <- data.frame(bee=c(0, 0, 1, 2, 3, 2, 0, 3),
                wasp=c(1, 3, 2, 0, 1, 1, 2, 1),
                  fly=c(1, 2, 4, 2, 1, 0, 1, 0),
              beetle=c(1, 0, 0, 1, 2, 2, 0, 2),
            butterfly=c(0, 0, 0, 0, 3, 1, 0, 1))

entropy(xx)

}
