% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/every_nth.R
\name{every_nth}
\alias{every_nth}
\alias{every_nth.default}
\alias{every_nth.matrix}
\alias{every_nth.data.frame}
\alias{every_nth.list}
\title{Select every n'th element}
\usage{
every_nth(x, ...)

\method{every_nth}{default}(x, n = 2, start = 1, ...)

\method{every_nth}{matrix}(x, n = 2, start = 1, margin = 1, ...)

\method{every_nth}{data.frame}(x, n = 2, start = 1, ...)

\method{every_nth}{list}(x, n = 2, start = 1, ...)
}
\arguments{
\item{x}{an object to be selected from}

\item{...}{further arguments passed to methods}

\item{n}{selection "step size"}

\item{start}{integer in [1:n] specifying the start of selection}

\item{margin}{for matrices, what margin to select along}
}
\description{
Select every second, third, fourth etc. element (or slice/hyperplane)
}
