% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{expand}
\alias{expand}
\title{Expand}
\usage{
expand(x, ...)
}
\arguments{
\item{x}{an object to be expanded}

\item{...}{further arguments passed to or from methods}
}
\value{
A vector with values and their repetitions specified by \code{x}
}
\description{
Expand a \code{"table"}, a \code{"table"}-like object, or a list of 
\code{"table"}-like objects
}
\examples{
x <- c(4, 2, 2, 2, 3, 3, 2, 4, 6, 6)
(xt <- table(x))
(xd <- dice(x))

expand(xt)
expand(xd)

expand(list(xt, xd, x))

xn <- as.table(1:4)
names(xn) <- LETTERS[1:length(xn)]
expand(xn)

}
\seealso{
\code{\link{dice}}, \code{\link{table}}
}
