% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode_obj.R
\name{explode_obj}
\alias{explode_obj}
\title{Explode object}
\usage{
explode_obj(x, indent = 2)
}
\arguments{
\item{x}{an R oject, or a character string describing an R object}

\item{indent}{how many spaces for indention (and exdention) at each level}
}
\value{
An exploded representation of the object is printed to console, and returned
invisibly. The output is in most cases a complete and reproducible
representation of the object, similarly to \code{dput}, but less compact and
more reaviling of its inner structure.
}
\description{
Presents an R object in an exploded form
}
\details{
If \code{x} is an R oject it is first deparsed and converted into a character
string describing the object. This string is then unwrapped, or exploded, 
according to these rules: newline and exdention after each open parenthesis,
newline and indention after each close parenthesis, and newline after each comma.
Parentheses and commas forming part of character strings are ignored.
}
\examples{
xc <- 'list(v=1, A=c("abv", "bom"), B=c(1:3, 31, 28), list("foo", "bar", 1))'
explode_obj(xc)

xl <- list(O=NA, R=list(j=1:3, h="(a)", q=structure(list(a=1:2, b=c("A, K",
  "B, L")), class="data.frame", row.names=c(NA, -2L))), N=1, L=FALSE)
explode_obj(xl)

mt <- 'coplot(mpg ~ disp | as.factor(cyl), data = mtcars,
       panel = panel.smooth, rows = 1)'
explode_obj(mt)
}
\seealso{
\code{\link{dput}}, \code{\link{dput2}}
}
