% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime.R
\name{factorise}
\alias{factorise}
\title{Factorise}
\usage{
factorise(x)
}
\arguments{
\item{x}{integer}
}
\value{
An integer vector
}
\description{
Find the prime factors of a given integer
}
\examples{
factorise(320)
factorise(2 * 2 * 2 * 3 * 3 * 5)

prod(factorise(5641324))

\dontrun{
factorise(nextn(60000000, c(2, 3)))
factorise(72*999983)
}
}
\seealso{
\code{\link{factors}} for unique prime factors or all integer factors
}
