% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime.R
\name{factors}
\alias{factors}
\title{Factors}
\usage{
factors(x, prime = FALSE)
}
\arguments{
\item{x}{an integer}

\item{prime}{should only prime factors be returned?}
}
\value{
An integer vector
}
\description{
Find the integers a given number is divisible by
}
\section{Note}{

The trivial factors \code{1} and \code{x} itself are not included.
}

\examples{
factors(210)
factors(210, prime=TRUE)

}
\seealso{
\code{\link{factorise}} for prime factorisation
}
