% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_name_ext.R
\name{file_ext}
\alias{file_ext}
\alias{file_name}
\alias{file_name_ext}
\title{File extension}
\usage{
file_ext(x)

file_name(x)

file_name_ext(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
\code{file_ext} returns the file extension of each file path. \code{file_name}
returns the file name, no extension, of each file path. \code{file_name_ext}
returns both name and extension, but arranged in separate columns of a matrix.
}
\description{
Separate file name and extension from a file path
}
\details{
If the supplied file name has several extensions, f.ex. like 
\code{foobar.tar.bz}, only the last extension will be considered.
}
\examples{
x <- c("/hg/.gi.tar.gz", "ff/hg/hh.pdf", "git", ".History", ".History.log")

file_ext(x)
file_name(x)
file_name_ext(x)
}
\seealso{
\code{\link{basename}}
}
