% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerknit.R
\name{fingerknit}
\alias{fingerknit}
\title{Render R example}
\usage{
fingerknit()
}
\value{
The clipboard is used for both input and output, but the output is also
returned invisibly as a character string. Warnings and errors are not captured,
but printed to console as normal. If an error is encountered nothing is returned
and the clipboard data remains unchanged.
}
\description{
Render example from R code stored on the clipboard
}
\details{
Similar to \code{reprex::reprex()} or \code{knitr::spin(text=, envir=new.env(), 
report=FALSE)}, but stripped down to the very basics. Input is plain valid R
code taken from the clipboard. It is run in a fresh environment and both 
commands and results are catured. Commands are kept as is, but results are 
commented out. Instead of using three backticks to indicate code for markdown, 
each line has four whitespaces prepended. \code{fingerknit} output is also
valid R code.
}
