% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitrange.R
\name{fitrange}
\alias{fitrange}
\title{Fit to a range}
\usage{
fitrange(W, lower = -1, upper = 1)
}
\arguments{
\item{W}{a numeric vector}

\item{lower}{the lower bound of the new vector}

\item{upper}{the upper bound of the new vector}
}
\description{
Linearly shift and scale a numeric vector so that it fits to a given range.
}
\examples{
range(fitrange(runif(9, -2, 0.1), 0, 1))

fitrange(c(2, 3, 5, 7, 4), 1, 0)
# same, but without warning
1 - fitrange(c(2, 3, 5, 7, 4), 0, 1)
}
\seealso{
\code{\link{norma}}
}
