% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten}
\alias{flatten}
\title{Flatten list}
\usage{
flatten(x, flatten.df = FALSE, keep.order = TRUE)
}
\arguments{
\item{x}{a list object}

\item{flatten.df}{should \code{data.frame}s also be flattened?}

\item{keep.order}{keep the order of the original list, same as seen when 
using \code{str}}
}
\value{
A single level list of \code{x}'s leaves.
}
\description{
Flatten a (nested) list to a list of its leaves
}
\details{
The nodes of the supplied list is traversed from root to leaf and successively
unlisted until no lists are left (except possibly for \code{data.frame}s).
}
\examples{
xl <- list(
  O=NA, 
  R=list(
    j=1:3,
    h="(a)",
    q=data.frame(
      a=1:2, 
      b=c("A, K", "B, L"),
      stringsAsFactors=FALSE
    )
  ), 
  N=1,
  L=FALSE
)

flatten(xl, flatten.df=TRUE, keep.order=FALSE)
flatten(xl, flatten.df=TRUE, keep.order=TRUE)
str(xl)
}
