% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_forecasts.R
\name{forecasts}
\alias{forecasts}
\title{Return forecasts}
\usage{
forecasts(x)
}
\arguments{
\item{x}{a \code{compare_forecasts} object}
}
\value{
A multivarite time series (\code{mts}) with the actual data, the holdout, 
on the first column, and the forecasts on the rest.
}
\description{
Return forecasts and actual data from \code{compare_forecasts} object
}
