% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime.R
\name{gcd}
\alias{gcd}
\title{Greatest common divisor}
\usage{
gcd(x, y)
}
\arguments{
\item{x, y}{integers whose greates common divisor is to be found}
}
\description{
Find the largest integer, that when two numbers are divided by it,
returns an integer in both cases
}
\examples{
gcd(sequence(10:16), rep(10:16, 10:16))

}
