% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intsect.R
\name{intsect}
\alias{intsect}
\title{Intersect}
\usage{
intsect(x)
}
\arguments{
\item{x}{list of sets (vectors of same mode or factors)}
}
\value{
A vector of same mode as input, or a single factor object if input was factor.
}
\description{
Performs set intersection on a list of vectors
}
\details{
The intersection between the sets in the list is found. This means no
duplicate values are returned, whether or not there were any in the
input.
}
\examples{
intsect(list(0:6, c(2, 4, 6, 8), 3:8))

fc <- factor(LETTERS[sample(1:5, 20, rep=TRUE)])
fcl <- split(fc, sample(1:3, 20, rep=TRUE))

intsect(fcl)
}
