% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime.R
\name{is_coprime}
\alias{is_coprime}
\title{Coprimality check}
\usage{
is_coprime(x, y)
}
\arguments{
\item{x, y}{integers to be tested for coprimality}
}
\value{
A logical vector
}
\description{
Test whether to integers are coprime, that is, have no factors in common
}
\examples{
is_coprime(sequence(10:16), rep(10:16, 10:16))
is_coprime(2*3*5*7, 11*13)

}
