% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime.R
\name{is_prime}
\alias{is_prime}
\title{Primality check}
\usage{
is_prime(x)
}
\arguments{
\item{x}{vector of integers}
}
\description{
Test integers for whether they are prime or not
}
\seealso{
\code{\link{primes}}
}
