% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime.R
\name{isPrime}
\alias{isPrime}
\title{Primality check}
\usage{
isPrime(x)
}
\arguments{
\item{x}{integer; one or more prime candidates}
}
\description{
Test an integer for whether it is prime or not
}
\seealso{
\code{\link{primes}}
}
