% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_finite.R
\name{keep_finite}
\alias{keep_finite}
\alias{keep_finite.default}
\alias{keep_finite.matrix}
\alias{keep_finite.data.frame}
\title{Keep finite values}
\usage{
keep_finite(x, ...)

\method{keep_finite}{default}(x, ...)

\method{keep_finite}{matrix}(x, margin = 1, keep = c("any", "complete"),
  ...)

\method{keep_finite}{data.frame}(x, margin = 1, keep = c("any", "complete"),
  ...)
}
\arguments{
\item{x}{a vector or matrix}

\item{...}{further arguments passed to methods}

\item{margin}{if \code{x} is matrix, which margin to keep finites by}

\item{keep}{if \code{x} is matrix, keep rows/columns with any finite
values, or keep only complete rows/columns.}
}
\value{
If \code{x} is a matrix and \code{margin} is 1 or 2, a matrix is returned.
Else a vector.
}
\description{
Remove \code{NA}s code{NaN}s and code{Inf}s from data
}
\examples{
m1 <- matrix(c(10, 20, 30, 43,
               10, NA, 32, 50,
               NA, NA, NA, NA,
               13, 22, 70, 81,
               NA, 29, NA, 41), 5, byrow=TRUE,
               dimnames=list(letters[1:5], LETTERS[1:4]))

keep_finite(m1) 
matplot(keep_finite(apply(m1, 2, sort, na.last=TRUE)), type="l")

m1[complete.cases(m1),]
keep_finite(m1, 1, "c") #same
keep_finite(m1, 2, "complete") #no complete columns

m1.df <- as.data.frame(t(m1))
keep_finite(m1.df, 2, "complete")
}
