% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.vector.R
\name{lag.vector}
\alias{lag.vector}
\title{Lag an arbitrary vector}
\usage{
\method{lag}{vector}(x, k, type = c("cycle", "na.fill", "trim"), ...)
}
\arguments{
\item{x}{vector to be lagged}

\item{k}{integer vector specifying the number of lags}

\item{type}{how to deal with non-overlapping sections}

\item{...}{further arguments passed to methods}
}
\description{
Lag an arbitrary vector
}
\examples{
x <- 1:9

lag.vector(x, c(0, 1, -2, 3))
lag.vector(x, c(0, 1, -2, 3), "na")
lag.vector(x, c(0, 1, -2, 3), "trim")
}
