% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_seq.R
\name{markov_seq}
\alias{markov_seq}
\title{Discrete markov sequence}
\usage{
markov_seq(tmat, init = 1, length = 1000)
}
\arguments{
\item{tmat}{a transition matrix}

\item{init}{the initial state}

\item{length}{length of the sequence}
}
\description{
Generate a random discrete markov sequence
}
\examples{
m <- matrix(c(0.5, 0.3, 0.2,
              0.2, 0.6, 0.2,
              0.2, 0.3, 0.5), 3, byrow=TRUE)

set.seed(1)
ms <- markov_seq(m)

colMeans(m)
prop.table(table(ms))
prop.table(table(tail(ms, -1), head(ms, -1), dnn=c("n", "n+1")), 1)
}
