% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_doc.R
\docType{data}
\name{math_constants}
\alias{math_constants}
\alias{e}
\alias{pi}
\alias{phi}
\alias{feig1}
\alias{feig2}
\alias{eu.ma}
\alias{khin}
\alias{glai.kin}
\title{Mathematical constants}
\format{An object of class \code{numeric} of length 1.}
\description{
Various mathemathical constants available as global variables
}
\details{
\describe{
  \item{\code{e}}{Euler's number}
  \item{\code{pi}}{Archimedes' number, the circle constant}
  \item{\code{phi}}{Golden ratio}
  \item{\code{feig1}}{Feigenbaum's first constant, \eqn{\delta}; 
bifurcation velocity}
  \item{\code{feig2}}{Feigenbaum's second constant, \eqn{\alpha}; 
reduction parameter}
  \item{\code{eu.ma}}{Euler–Mascheroni constant}
  \item{\code{khin}}{Khintchine's constant}
  \item{\code{glai.kin}}{Glaisher-Kinkelin constant}
}
}
\keyword{datasets}
