% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_doc.R
\docType{data}
\name{mathemathical_constants_char}
\alias{mathemathical_constants_char}
\alias{e.char}
\alias{pi.char}
\alias{phi.char}
\alias{feig1.char}
\alias{feig2.char}
\alias{eu.ma.char}
\alias{khin.char}
\alias{glai.kin.char}
\title{High precision mathematical constants}
\format{An object of class \code{character} of length 1.}
\description{
Character strings representing various mathemathical constants to ~100 
decimal points
}
\details{
\describe{
  \item{\code{e.char}}{Euler's number}
  \item{\code{pi.char}}{Archimedes' number, the circle constant}
  \item{\code{phi.char}}{Golden ratio}
  \item{\code{feig1.char}}{Feigenbaum's first constant, \eqn{\delta}; 
bifurcation velocity}
  \item{\code{feig2.char}}{Feigenbaum's second constant, \eqn{\alpha}; 
reduction parameter}
  \item{\code{eu.ma.char}}{Euler–Mascheroni constant}
  \item{\code{khin.char}}{Khintchine's constant}
  \item{\code{glai.kin.char}}{Glaisher-Kinkelin constant}
}
}
\keyword{datasets}
