% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_multiple.R
\name{merge_multiple}
\alias{merge_multiple}
\title{Merge multiple \code{data.frame}s}
\usage{
merge_multiple(x, by, all = FALSE, sort = TRUE, incomparables = NULL,
  include = NULL, exclude = NULL)
}
\arguments{
\item{x}{a list of \code{data.frame}s with at least one column in common}

\item{by}{name of the column to be merged by, by default the full intersect of 
column names between \code{data.frame}s}

\item{all}{include all rows, including those with no match}

\item{sort}{sort the output on the \code{by} column(s)}

\item{incomparables}{values which cannot be matched.}

\item{include, exclude}{numric, logical or character vector specifying which 
columns to include in or exclude from the merge}
}
\description{
Merge multiple \code{data.frame}s
}
\details{
If there are duplicate columns that aren't being used to merge by, one of two
things will happen. If the parent \code{data.frame}s of the duplicate columns
are named, then that name will be appended to the relevant column names. If
the \code{data.frame}s aren't named, then the \code{data.frame}s idices in
the parent list are appended to the relevant column names.\cr
Inclusion and exclusion are performed in sequence, so that if both 
\code{include} and \code{exclude} are specified, \code{exclude} acts on the
result from \code{include}.
}
\examples{
dtf1 <- data.frame(ast=1:4, bar=1:4, kat=c("A", "B", "C", "D"))
dtf2 <- data.frame(ast=1:6, bar=1:6, jun=9:4)
dtf3 <- data.frame(ast=2:6, bar=2:6, kat=c("A", "B", "C", "D", "E"))
dtf4 <- data.frame(ast=3:4, bar=3:4)
dtf5 <- data.frame(ast=1:-3, bar=0:4, git=0:4)

ll <- list(d1=dtf1, d2=dtf2, dtf3, A=dtf4, dtf5)

merge_multiple(ll, by="bar")
merge_multiple(ll, by="bar", all=TRUE, include=1:2)
merge_multiple(ll, by="bar", all=TRUE, exclude="kat")
merge_multiple(x=ll, by=c("bar", "ast"), all=TRUE)
}
