% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidensity.R
\name{multidensity}
\alias{multidensity}
\title{Plot multiple kernel density estimates}
\usage{
multidensity(x, main, xlab = "", ylab = "Density", xlim, ylim,
  col = 1:9, lty = 1:2, lwd = 1, add = FALSE, frame.plot = TRUE,
  legend = TRUE, x.legend = "topleft", y.legend = NULL, bty = "o",
  box.col = "#FFFFFF00", bg.legend = "#FFFFFFAA", cex.legend = 0.7,
  x.intersp = 1, y.intersp = 1.5, inset = 0, xpd.legend = NA,
  horiz = FALSE, ...)
}
\arguments{
\item{x}{a list or data.frame of numeric values}

\item{main}{a main title for the plot. Defaults to the call made
to \code{density}}

\item{xlab, ylab}{labels for the x and y axes}

\item{xlim, ylim}{the x and y limits of the plot}

\item{col, lty, lwd}{the line colours, types and widths for lines appearing
in plot and legend}

\item{add}{if \code{TRUE}, add to the current plot}

\item{frame.plot}{an integer indicating whether a box should be drawn around
the plot before the legend (\code{1}), after the legend (\code{2}),
or not at all (\code{0}). Logical values are coerced to integer, so
\code{TRUE} implies \code{1}, and \code{FALSE} implies \code{0}}

\item{legend}{logical; if \code{TRUE} (the default) a legend is included with 
the plot}

\item{x.legend, y.legend}{the x and y co-ordinates to be used to position the
legend. They can be specified by keyword or in any way which is 
accepted by xy.coords}

\item{bty}{legend box type}

\item{box.col}{line colour for the legend box}

\item{bg.legend}{background colour for the legend box}

\item{cex.legend}{character expanson faftor for legend}

\item{x.intersp, y.intersp}{horizontal and vertical character interspacing for
legend}

\item{inset}{the legends inset distance from the margins as a fraction of 
the plot region}

\item{xpd.legend}{the value of \code{xpd} to be used while drawing the legend}

\item{horiz}{logical; if \code{TRUE}, set the legend horizontally rather than
vertically}

\item{...}{further arguments passed to \code{density}}
}
\value{
An invisible list of the \code{"density"} objects the plot is based on.
}
\description{
Plot multiple kernel density estimates in the same window, together with a
legend
}
\examples{
set.seed(1)
dl <- list("Unif-1"=runif(80, -2.1, 2.1), 
           "Unif-2"=runif(70, -1.5, 1.5), 
         "Normal-1"=rnorm(50, 0, 0.866), 
         "Normal-2"=rnorm(90, 0, 1))

# sqrt((sd^2)*12) # sd to unif range

md <- multidensity(dl)
head(md, 2)

multidensity(dl, adj=1.2, x.leg="topright", frame=FALSE, inset=-0.02, lty=1)
multidensity(dl, x.legend="top", horiz=TRUE, cex.legend=0.5,
  inset=-0.05, bg.legend="white")
}
\seealso{
\code{\link{density}}, \code{\link{ahist}}
}
