% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narm.R
\name{narm}
\alias{narm}
\alias{narm.default}
\alias{narm.matrix}
\alias{narm.data.frame}
\title{Remove \code{NA}s}
\usage{
narm(x, ...)

\method{narm}{default}(x, ...)

\method{narm}{matrix}(x, margin = 1, keep = c("any", "complete"), ...)

\method{narm}{data.frame}(x, margin = 1, keep = c("any", "complete"), ...)
}
\arguments{
\item{x}{a vector or matrix}

\item{...}{further arguments passed to methods}

\item{margin}{if \code{x} is matrix, which margin to remove \code{NA}s by}

\item{keep}{if \code{x} is matrix, keep rows/columns with any non-\code{NA}
values, or keep only complete rows/columns.}
}
\value{
If \code{x} is a matrix and \code{margin} is 1 or 2, a matrix is returned.
Else a vector.
}
\description{
Remove \code{NA}s from vector or matrix
}
\examples{
m1 <- matrix(c(10, 20, 30, 43,
               10, NA, 32, 50,
               NA, NA, NA, NA,
               13, 22, 70, 81,
               NA, 29, NA, 41), 5, byrow=TRUE,
               dimnames=list(letters[1:5], LETTERS[1:4]))

narm(m1) 
matplot(narm(apply(m1, 2, sort, na.last=TRUE)), type="l")

m1[complete.cases(m1),]
narm(m1, 1, "c") #same
narm(m1, 2, "complete") #no complete columns

m1.df <- as.data.frame(t(m1))
narm(m1.df, 2, "complete")
}
