% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitrange.R
\name{norma}
\alias{norma}
\title{Normalize}
\usage{
norma(W, c = 0, r = 2)
}
\arguments{
\item{W}{a numeric vector}

\item{c}{the centre (as in the midrange) for the new vector}

\item{r}{the range of the new vector}
}
\description{
Linearly shift and scale a numeric vector so that it has a given range, about
a given centre.
}
\examples{
range(norma(runif(9, -2, 0.1), 0, 2))
}
\seealso{
\code{\link{fitrange}}
}
