% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.histogram.R
\name{plot.histogram}
\alias{plot.histogram}
\title{Plot \code{histogram} object}
\description{
A a very minor modification of \code{graphics::plot.histogram}. \cr
Only difference is that \code{lwd} now specifies the width of the histogram bars'
outline.
}
\seealso{
\code{\link[graphics]{plot.histogram}}, \code{\link{plot.stl}}
}
