% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.stl.R
\name{plot.stl}
\alias{plot.stl}
\title{Plot \code{stl} object}
\usage{
\method{plot}{stl}(x, labels = colnames(X), set.pars = list(mar = c(0, 6, 0,
  6), oma = c(6, 0, 4, 0), tck = -0.01, mfrow = c(nplot, 1)), main = NULL,
  range.bars = TRUE, ..., col.range = "light gray")
}
\description{
A a very minor modification of \code{stats::stl}. \cr
Only difference is that the distance between the plotting window and
the x and y labels is set by \code{par("mgp")[1]}, as it is for regular plots.
}
\seealso{
\code{\link[stats]{plot.stl}}, \code{\link{plot.histogram}}
}
