% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime.R
\name{primes}
\alias{primes}
\title{Prime number generator}
\source{
\url{https://stackoverflow.com/questions/3789968/generate-a-list-of-primes-up-to-
a-certain-number/3791284#3791284}
}
\usage{
primes(n)
}
\arguments{
\item{n}{integer; all prime numbers up to this will be returned}
}
\description{
Prime generator based on the sieve of Eratosthenes
}
\details{
Effective for primes up to ~100,000,000. \cr
On my lightweight laptop: 1e7 -> 0.32s, 5e7 -> 1.7s, 1e8 -> 3.7s, 2e8 -> 7.6s, 3e8 -> 15s
}
\seealso{
\code{\link{isPrime}}
}
