% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartz.png.R
\name{quartz.png}
\alias{quartz.png}
\title{Save as PNG}
\usage{
quartz.png(file = "\%Y\%m\%d_\%H", width = 600, dir, force = FALSE)
}
\arguments{
\item{file}{file name. If it contains any \code{"\%"} it is passed on as a 
format string to \code{format(Sys.time(), file)}. A \code{.png} file 
extension is added automatically.}

\item{width}{pixel width of the PNG file}

\item{dir}{directory to save to. Defaults to current working directory}

\item{force}{force overwriting of existing file with same name. By default
duplicate path names are resolved by appending \code{_N}, using 
successive integers, to the end of the file name.}
}
\value{
A PNG file is written to disk and a message is written to the console, giving
the new file's path and pixel dimensions. The file path is also returned
invisibly.
}
\description{
Save the contents of the current Quartz window as PNG file
}
\examples{
\dontrun{
set.seed(1)
i_h100 <- round(runif(100, 2, 30), 2)
i_cd <- rexp(100, 1/i_h100)
mydata <- data.frame(i_cd, i_h100)

mydata$i_h100_2m <- cut(mydata$i_h100, seq(2, 30, by=2))

i_cd_2m <- aggregate(i_cd ~ i_h100_2m, mydata, mean)

set_mar(x=2.5)
plot.default(i_cd_2m, xaxt="n", main="Groupwise means", xlab="", cex.main=1)
axis(1, i_cd_2m[,1], as.character(i_cd_2m[,1]), cex.axis=0.6, las=2)

quartz.png()
p <- quartz.png("test", 550)
file.info(p)
}
}
