% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle2.R
\name{quick_table}
\alias{quick_table}
\title{Tabulate data}
\usage{
quick_table(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector or factor object}

\item{na.rm}{should \code{NA}s be included}
}
\value{
A \code{data.frame} with columns \code{val} (the original values and class
of \code{x}) and \code{freq} (the count, or frequency, of each value in
\code{x}, integer). The rows are sorted by frequency in descending order.
}
\description{
Quick and simple function for creating contingency tables
}
\examples{
set.seed(1)
m <- sample(c(rep(NA, 5), rpois(45, 3)))
quick_table(m)
}
