% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randcolours.R
\name{randcolours}
\alias{randcolours}
\title{Random colours}
\usage{
randcolours(n, l = c(0.2, 0.9), c1 = c(0, 1), c2 = c(0, 1),
  alpha = 1, space = c("Luv", "Lab"))
}
\arguments{
\item{n}{number of colours}

\item{l}{lightness range}

\item{c1}{colour channel one range}

\item{c2}{colour channel two range}

\item{alpha}{alpha channel range}

\item{space}{should the parameters be interpreted as Luv or Lab components?}
}
\description{
Generate a randomly selected colour palette
}
\details{
The range of \code{l}, \code{c1}, \code{c2} and \code{alpha}, will be 
interpreted as the wanted range of each colour component, whether their length
is 1, 2, or more. Although they all should nominally lie within [0, 1], only 
\code{alpha} must do so to achieve a valid output. The others can exeed this
range, at an icreased risk of clipping.
}
\examples{
set.seed(3)
n <- 20
plot(1:n, col=randcolours(n), pch=16, cex=5)
}
