% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar_and_par.R
\name{revert_par}
\alias{revert_par}
\title{Revert par}
\usage{
revert_par()
}
\description{
Reverts \code{par} settings back to \code{old.par}
}
\seealso{
Other par_and_plot_margins_functions: \code{\link{default_par}},
  \code{\link{set_mar}}
}
