% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle2.R
\name{rle2}
\alias{rle2}
\title{Run Length Encoding}
\usage{
rle2(x, na.unique = FALSE, output = c("data.frame", "rle",
  "named vector", "lengths", "values"))
}
\arguments{
\item{x}{a numeric or character vector}

\item{na.unique}{should every \code{NA} be conidered unique?}

\item{output}{what form of output}
}
\value{
Return value depends on \code{output}. 
\describe{
  \item{\code{data.frame}}{A data.frame with lengths and values columns}
  \item{\code{rle}}{An object of class \code{"rle"}}
  \item{\code{named vector}}{A vector of lengths with values as names}
  \item{\code{lengths}}{The lengths as a single vector}
  \item{\code{values}}{The values as a single vector}
}
}
\description{
Compute the lengths and values of runs of equal values in a vector
}
\examples{

x <- c(NA, NA, 1, 2, 3, 3, NA, NA, NA, 2, 2, 2, NA, 1, 1, NA, NA)
rle2(x)

m <- matrix(c(
  0.7, 0.2, 0.1,
  0.2, 0.6, 0.2,
  0.1, 0.2, 0.7
), 3, byrow=TRUE)

set.seed(1)
y <- LETTERS[markov_seq(n=100, m)]
rle2(y, out="named")


# Same result as rle
rle2(x, na.unique=TRUE, output="rle")
rle(x)

# inverse.rle works as long as output is "rle"
inverse.rle(rle2(x, output="rle"))
}
