% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_range.R
\name{seq_range}
\alias{seq_range}
\title{Generate a sequence spanning a given range}
\usage{
seq_range(x, ..., spread)
}
\arguments{
\item{x}{a single numeric, a range, or a sequence}

\item{...}{further arguments passed to \code{seq}}

\item{spread}{use \code{spread_seq} to spread out the range by a given factor}
}
\description{
Generate a sequence spanning a given range
}
\details{
If \code{x} is a single number, the range is interpreted to be \code{[0, x]}.
If \code{x} is length two, the numbers are interpreted as the left and right
extrema of the sequence interval.
If \code{x} is longer than two, the sequence is based upon its \code{range}.
}
\examples{

seq_range(c(1, 4), by=0.5)
seq_range(c(1, 4), by=0.5, spread=2)
seq_range(4)

x <- sample(1:10, 3)
seq_range(x)
}
