% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{smat}
\alias{smat}
\title{Similarity matrix}
\usage{
smat(x, y, s, byrow = FALSE)
}
\arguments{
\item{x}{an object containing the values the similarity matrix should be
computed for}

\item{y}{same as \code{x}. If given the union of values in \code{x} and
\code{y} are used, if not the unique values of \code{x} are used}

\item{s}{a vector for filling the matrix. By default producing an identity
matrix}

\item{byrow}{should s fill the matrix by row?}
}
\value{
A square matrix with the values of \code{s} and row-/colnames of the unique
values in \{x, y\}.
}
\description{
Create a similarity matrix
}
\examples{
smat(1:3)

smat(c("f", "e", "d"), s=c(
	4, 1, 1,
	1, 3, 2,
	1, 2, 3
))

}
\seealso{
\code{\link{similarity}}
}
