% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_doc.R
\docType{data}
\name{speedskate}
\alias{speedskate}
\title{2018 MarbleLympics speed skating times}
\format{A list containing two data.frames, one for each lane. Columns are heat and rows
are time checks in seconds.}
\source{
https://www.youtube.com/watch?v=fA-O6f_jArk
}
\usage{
speedskate
}
\description{
Intermediate and total times for all 16 runs, arranged by lane and heat
number.
}
\examples{
tt <- t(do.call(cbind, speedskate))
pairs(tt)
cor(tt)
outer(
  colnames(tt), 
  colnames(tt), 
  Vectorize(function(i,j) cor.test(tt[,i],tt[,j])$p.value)
)
}
\keyword{datasets}
