% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_range.R
\name{spread_seq}
\alias{spread_seq}
\title{Spread sequence}
\usage{
spread_seq(x, f = 1.1, ..., node = c("midrange", "first", "last", "mean",
  "median", "min", "max"))
}
\arguments{
\item{x}{a numeric vector}

\item{f}{numeric or item matching a function. If numeric, a multiplicative
factor applied to the distance between points, otherwise a function to be
applied to differences}

\item{...}{further arguments passed to \code{f} if matching a function}

\item{node}{the location of \code{x} that will remain unchanged}
}
\description{
Spread out the values of a numeric vector
}
\examples{
x <- c(-1, 0, 2, NA, 4, 5, 6, 8, 9)

spread_seq(c(-1, 1))
spread_seq(x, 1.5, "midrange")
spread_seq(x, 1.5, "first")
spread_seq(x, 1.5, "last")
spread_seq(x, 1.5, "mean")
spread_seq(x, 1.5, "median")

spread_seq(c(3, 4, 1, 9, 6))
spread_seq(c(3, 4, 1, 9, 6), 2, "min")
spread_seq(c(3, 4, 1, 9, 6), 2, "max")

spread_seq(c(3, 4, 1, 9, 6), "/", 2)
spread_seq(c(3, 4, 1, 9, 6), 0.5)

y <- sort(c(3, 4, 1, 9, 6))
plot(y)
lines(spread_seq(y, "log1p"))

f <- function(x) sqrt(abs(x)*2)*sign(x)
y <- c(3, 4, 1, 9, 6)
plot(y)
lines(spread_seq(y, f=f))
}
