% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.stl.R
\name{summary.stl}
\alias{summary.stl}
\title{Summarizing seasonal decomposition}
\usage{
\method{summary}{stl}(object, digits = getOption("digits"), ...)
}
\arguments{
\item{object}{an object of class "\code{stl}"}

\item{digits}{the number of significant digits to use when printing}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for class "\code{stl}".
}
\details{
This function is a slight modification to \code{stats:::summary.stl}, the main
change being the addition of the variance statistic, which can be considered
a parametric (normal) compliment to the existing IQR statistic.
}
\examples{
set.seed(1)
x <- ts(rnorm(1e4, sd=1), frequency=12)
a <- stl(x, s.window="periodic")
stats:::summary.stl(a)
summary(a)

b <- stl(co2, s.window="periodic")
summary(b)
}
