% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_doc_package.R
\docType{package}
\name{ymse}
\alias{ymse}
\title{ymse: A collection of more or less useful functions}
\description{
There is no grand "theme" to ymse, other than that none of the functions,
and in some cases function groups and classes, seemed to fit too well in any
other package or merit their own package entirely.
}
\section{ymse functions}{

\code{\link{addrows}} Add rown to a data.frame
\code{\link{ahist}} Create an average shifted histogram
}

