% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central_tendency.R
\name{central.tendency}
\alias{central.tendency}
\alias{pseudomedian}
\alias{cmode}
\title{Central tendency measures}
\usage{
pseudomedian(x)

cmode(x, ...)
}
\arguments{
\item{x}{numeric vector}

\item{...}{send further arguments to underlying function, e.g. \code{density}
for cmode}
}
\description{
Central tendency measures
}
\examples{
xx <- c(1, 3, 4, 5, 7, 8, 9, 9, 7, 5, 4, 5, 3, 8)
median(xx)#' 
pseudomedian(xx)

}
\seealso{
\code{\link{means}}
}
