% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_seq.R
\name{markov_seq}
\alias{markov_seq}
\title{Discrete markov sequence}
\usage{
markov_seq(n = 100, tmat = rbind(1:3, 3:1, 2:0), init = 1)
}
\arguments{
\item{n}{length of the sequence}

\item{tmat}{a transition matrix}

\item{init}{the initial state}
}
\description{
Generate a random discrete markov sequence
}
\examples{
m <- matrix(c(0.5, 0.3, 0.2,
              0.2, 0.6, 0.2,
              0.2, 0.3, 0.5), 3, byrow=TRUE)

set.seed(1)
ms <- markov_seq(n=1000, tmat=m)

colMeans(m)
prop.table(table(ms))
round(prop.table(table(head(ms, -1), tail(ms, -1), dnn=c("n", "n+1")), 1), 2)
}
