% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.stl.R
\name{plot.stl}
\alias{plot.stl}
\title{Plot \code{stl} object}
\description{
A a very minor modification of \code{stats::stl}. \cr
Only difference is that the distance between the plotting window and
the x and y labels is set by \code{par("mgp")[1]}, as it is for regular plots.
}
\seealso{
\code{\link[stats]{plot.stl}}, \code{\link{plot.histogram}}
}
