% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle2.R
\name{quick_table}
\alias{quick_table}
\title{Tabulate data}
\usage{
quick_table(x, na.rm = FALSE, order = c("frequency", "value", "none"))
}
\arguments{
\item{x}{a vector or factor object}

\item{na.rm}{should \code{NA}s be included}

\item{order}{how should the results be ordered, if any?}
}
\value{
A \code{data.frame} with columns \code{val} (the original values and class
of \code{x}) and \code{freq} (the count, or frequency, of each value in
\code{x}, integer). The rows are sorted by frequency in descending order.
}
\description{
Quick and simple function for creating contingency tables
}
\examples{
set.seed(1)
m <- sample(c(rep(NA, 5), rpois(45, 3)))
quick_table(m)

x <- LETTERS[c(2, 2, 2, 2, 3, 1, 1)]
quick_table(x, order="freq")
quick_table(x, order="value")
quick_table(x, order="none")
}
