% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_package.R
\docType{package}
\name{ymse}
\alias{ymse}
\alias{ymse-package}
\title{ymse: A collection of more or less useful functions}
\description{
There is go grand "theme" to ymse, other than that none of the functions,
and in some cases function groups, seemed to fit too well in any other package
}
\section{ymse functions}{

addrows
}

